+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - none
 
 # v2023.4.1-beta
-- add google-colab support
+- sh - add cloudflared-tunnel support
+- sh - add google-colab support
 
 # v2023.3.21
 - ci - update shell-function shRawLibFetch()